requirements.txt
# 마법옷장 FastAPI 애플리케이션 의존성

# 웹 프레임워크
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 템플릿 엔진
jinja2>=3.1.0

# 데이터베이스 & 인증
asyncpg>=0.29.0
supabase>=2.3.0

# JWT 및 보안
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
cryptography>=42.0.0

# 데이터 검증 및 직렬화
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 환경 변수 관리
python-dotenv>=1.0.0
aiofiles>=23.2.0

# HTTP 클라이언트
httpx>=0.25.0

# 보안 라이브러리
bleach>=6.1.0
itsdangerous>=2.1.0

# 파일 업로드 (Wasabi/S3)
boto3>=1.34.0
botocore>=1.34.0

# 이미지 처리
pillow>=10.1.0

# 로깅 및 모니터링
structlog>=23.2.0
loguru>=0.7.0

# 웹소켓
websockets>=12.0

# 유틸리티
python-slugify>=8.0.0
email-validator>=2.1.0

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# 코드 품질
black==23.11.0
isort==5.12.0
flake8==6.1.0

# 타입 체크
mypy==1.7.1

# 테스트 커버리지
pytest-cov==4.1.0